# backend/public/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Transaction Dashboard API",
        "description": "API Documentation for Transaction Dashboard",
        "version": "1.0"
    },
    "paths": {
...
        "/dashboard": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get all dashboard data",
                "description": "Returns all dashboard analytics data",
                "operationId": "faae45987996934eda46ad27a6f61dea",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Start date (Y-m-d H:i:s)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "End date (Y-m-d H:i:s)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Period: daily, weekly, monthly",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "daily",
                                "weekly",
                                "monthly"
                            ]
                        }
                    },
                    {
                        "name": "user_type",
                        "in": "query",
                        "description": "User type: domestic, foreign",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "domestic",
                                "foreign"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
        "/dashboard/trends": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get trends data",
                "description": "Returns transaction trends",
                "operationId": "45f3f41f3044dd78ced5561e5dfdc533",
                "parameters": [
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Period: daily, weekly, monthly",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "daily",
                                "weekly",
                                "monthly"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/dashboard/trending-items": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get trending items",
                "description": "Returns trending items",
                "operationId": "5eba929716106511b1959f1b4d117669",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/dashboard/top-buyers": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get top buyers",
                "description": "Returns top buyers",
                "operationId": "92ff0c936963af33d7c75066d5f535b1",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of buyers",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/dashboard/top-sellers": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get top sellers",
                "description": "Returns top sellers",
                "operationId": "2242ecb2d7521befa4a165c808e412b2",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of sellers",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
        "/dashboard/top-items": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get top items by revenue",
                "description": "Returns top items performance by revenue",
                "operationId": "765cbca218f5430fa61e56f9f62aa645",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items to return",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Start date (Y-m-d H:i:s)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "End date (Y-m-d H:i:s)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_type",
                        "in": "query",
                        "description": "User type: domestic, foreign",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "domestic",
                                "foreign"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/dashboard/price-distribution": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get price distribution",
                "description": "Returns price distribution histogram",
                "operationId": "770bec0a0ab9e7fa3665c02be4c1129b",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Start date (Y-m-d H:i:s)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "End date (Y-m-d H:i:s)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_type",
                        "in": "query",
                        "description": "User type: domestic, foreign",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "domestic",
                                "foreign"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme.",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}